custom-elements/registering-custom-elements/unresolved-elements-namespace-test-001.html

<!DOCTYPE html>
<html>
<head>
<title>If prototype interface inherits from SVGElement then namespace must be SVG Namespace</title>
<meta name="author" title="Vasiliy Degtyarev" href="mailto:vasya@unipro.ru">
<meta name="assert" content="If prototype's interface inherits from SVGElement, then NAMESPACE must be SVG Namespace">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    var xsvg = doc.createElementNS(SVG_NAMESPACE, 'x-svg');
    assert_equals(xsvg.namespaceURI, SVG_NAMESPACE, 'Wrong SVG element namespace');

    var proto = Object.create(xsvg.constructor.prototype);
    var GeneratedConstructor = doc.register('x-svg-1', {prototype: proto});
    var xsvg1 = new GeneratedConstructor();
    
    assert_equals(xsvg1.namespaceURI, SVG_NAMESPACE, 'Wrong namespace');
    
}, 'Namespace should be SVG for SVGElement');

</script>
</body>
</html>
